backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-generativeai
chromadb
sentence-transformers
numpy